requirements.txt
treamlit_option_menu
beautifulsoup4
streamlit
lxml